refactor(db): extract book read path to db/src/books.rs
