+++ db/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod app_state;
 pub mod auth;
 pub mod author_photos;
+pub mod books;
 pub mod covers;
 pub mod ebook;
 pub mod helpers;
@@ -25,6 +26,11 @@ pub mod worker;
 // instead of `omnibus_db::queries::list_books(...)`. Keeps callsites terse
 // and mirrors how `db.rs` looked before the extraction.
 pub use app_state::*;
+pub use books::{
+    count_books, count_search_books, get_book, get_book_by_uuid, library_from_db,
+    library_from_db_with_total, list_books, list_indexed_rows, resolve_book_id_by_uuid,
+    search_books, IndexedRow, MAX_BOOKS_RETURNED,
+};
 pub use covers::{covers_dir, get_cover, get_last_modified_epoch};
 pub use helpers::{build_fts_match, sanitize_fts_query};
 pub use metadata_overrides::*;
